fix widget-image issue

+++ src/app/views/widget/widget-edit/widget-image/widget-image.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {WidgetService} from '../../../../service/widget.service.client';
 import {ActivatedRoute, Router} from '@angular/router';
 import {Widget} from '../../../../models/widget.model.client';
+import {NgForm} from '@angular/forms';
 
 
 @Component({
@@ -10,36 +11,74 @@ import {Widget} from '../../../../models/widget.model.client';
   styleUrls: ['./widget-image.component.css']
 })
 export class WidgetImageComponent implements OnInit {
-
+  @ViewChild('f') imageForm: NgForm;
   userId: String;
   websiteId: String;
   pageId: String;
   widgetId: String;
   widget: Widget;
+  text: String;
+  url: String;
+  width: String;
 
   constructor(private widgetService: WidgetService,
               private activatedRoute: ActivatedRoute,
               private router: Router) {
   }
 
+  delete() {
+    if (this.widgetId !== undefined) {
+      this.widgetService.deleteWidget(this.widgetId);
+      this.router.navigate(['../'], {relativeTo: this.activatedRoute});
+    } else {
+      this.router.navigate(['../../'], {relativeTo: this.activatedRoute});
+    }
+  }
+
+  updateOrCreate() {
+    this.widget.text = this.imageForm.value.imageText;
+    if (this.imageForm.value.imageUrl !== '') {
+      this.widget.url = this.imageForm.value.imageUrl;
+    } else if (this.widgetId !== undefined) {
+      this.router.navigate(['../'], {relativeTo: this.activatedRoute});
+      return;
+    } else {
+      this.router.navigate(['../../'], {relativeTo: this.activatedRoute});
+      return;
+    }
+    this.widget.width = this.imageForm.value.imageWidth;
+    this.widget.widgetType = 'IMAGE';
+    if (this.widgetId !== undefined) {
+      this.widgetService.updateWidget(this.widgetId, this.widget);
+      this.router.navigate(['../'], {relativeTo: this.activatedRoute});
+    } else {
+      this.widgetService.createWidget(this.pageId, this.widget);
+      this.router.navigate(['../../'], {relativeTo: this.activatedRoute});
+    }
+  }
+
   ngOnInit() {
-    this.activatedRoute.params.subscribe(
-      (params: any) => {
-        this.widgetId = params['wgid'];
-        this.pageId = params['pid'];
-        this.userId = params['uid'];
-        this.websiteId = params['wid'];
+    this.activatedRoute.params.subscribe(params => {
+      this.pageId = params['pid'];
+      this.widgetId = params['wgid'];
+      if (params['wgid'] !== undefined) {
+        this.widget = this.widgetService.findWidgetById(this.widgetId);
+      } else {
+        this.widget = new Widget('', '', '', '', '', '', '');
       }
-    );
-    this.widget._id = this.widgetId;
-    this.widgetService.findWidgetById(this.widgetId);
+      this.text = this.widget.text;
+      this.url = this.widget.url;
+      this.width = this.widget.width;
+    });
   }
-  updateWidgetController(widget: Widget) {
-    this.widgetService.updateWidget(widget._id, widget);
-    this.router.navigate([this.newWidget.url]);
-  }
-  deleteWidgetController(widget: Widget) {
-    this.widgetService.deleteWidget(this.widgetId);
-    this.router.navigate([this.newWidget.url]);
-  }
+
+
+  // updateWidgetController(widget: Widget) {
+  //   this.widgetService.updateWidget(widget._id, widget);
+  //   this.router.navigate([this.newWidget.url]);
+  // }
+  // deleteWidgetController(widget: Widget) {
+  //   this.widgetService.deleteWidget(this.widgetId);
+  //   this.router.navigate([this.newWidget.url]);
+  // }
 }
